Delete navigation when sector is deleted
