+++ package.json
@@ -32,6 +32,7 @@
     "history": "^4.7.2",
     "is-url-external": "^1.0.3",
     "lodash": "^4.17.4",
+    "moment": "^2.22.1",
     "node-sass-chokidar": "1.1.0",
     "npm-run-all": "^4.1.1",
     "pluralize": "^7.0.0",
@@ -67,7 +68,7 @@
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-prettier": "^2.3.1",
     "eslint-plugin-react": "^7.5.1",
-    "firebase-tools": "^3.18.4",
+    "firebase-tools": "^3.18.5",
     "prettier": "^1.7.4",
     "react-scripts": "^1.0.1",
     "react-test-renderer": "^16.0.0",
